+++ files/RELEASES.md
@@ -585,3 +585,13 @@ Based on:
 - [python v0.34.0] files
 ### Releases
 - [PyPI v0.34.0] https://pypi.org/project/codat-files/0.34.0 - files
+
+## 2023-08-19 00:01:05
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.74.3 (2.86.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.34.1] files
+### Releases
+- [PyPI v0.34.1] https://pypi.org/project/codat-files/0.34.1 - files